// schemas/admin.ts
export const userGroupSchema = z.object({
  userGroupId: z
    .number()
    .min(0, { message: "User group ID must be 0 or greater" }),
  userGroupCode: z
    .string()
    .min(5, { message: "User group code must be at least 5 characters" })
    .max(50, { message: "User group code must be less than 50 characters" }),
  userGroupName: z
    .string()
    .min(5, { message: "User group name must be at least 5 characters" })
    .max(100, { message: "User group name must be less than 10 characters" }),
  remarks: z
    .string()
    .max(255, { message: "Remarks cannot exceed 255 characters" })
    .optional(),
  isActive: z.boolean(),
})

export type UserGroupSchemaType = z.infer<typeof userGroupSchema>